api/requirements.txt
CacheControl==0.12.5
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
firebase-admin==2.15.0
Flask==1.0.2
Flask-Cors==3.0.7
flask-swagger-ui==3.18.0
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
grpcio==1.18.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
msgpack==0.6.0
Pillow==5.4.1
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
